refactor(core): move db and git modules under src/core
