+++ scripts/profile-review-bundles.mjs
@@ -8,9 +8,9 @@
 import fs from "node:fs";
 import path from "node:path";
 import os from "node:os";
-import { openDb } from "../db.js";
+import { openDb } from "../src/core/db.js";
 import { syncAll } from "../src/sync/sync.js";
-import { isGitRepository, getHeadCommitHash } from "../git.js";
+import { isGitRepository, getHeadCommitHash } from "../src/core/git.js";
 import {
   buildReviewBundlePlan,
   createReviewBundleArtifacts,
