+++ src/index.js
@@ -6,9 +6,9 @@ import fs from "node:fs";
 import path from "node:path";
 import { randomUUID } from "node:crypto";
 import { fileURLToPath } from "node:url";
-import { openDb, checkpointJobFinish, loadStalledJobs, pruneJobCheckpoints } from "../db.js";
+import { openDb, checkpointJobFinish, loadStalledJobs, pruneJobCheckpoints } from "./core/db.js";
 import { syncAll, isSyncDirWritable, getSyncOwnershipDiagnostics, isStructuralProjectId } from "./sync/sync.js";
-import { isGitAvailable, isGitRepository, initGitRepository, getSceneProseAtCommit } from "../git.js";
+import { isGitAvailable, isGitRepository, initGitRepository, getSceneProseAtCommit } from "./core/git.js";
 import { createAsyncJobManager, readJsonIfExists } from "./runtime/async-jobs.js";
 import {
   createHelpers,
